requirements.txt
cassandra-driver==3.25.0
certifi==2021.10.8
click==8.1.3
colorama==0.4.4
Flask==2.1.2
geomet==0.2.1.post1
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
numpy==1.22.4
pandas==1.4.2
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
Werkzeug==2.1.2
wincertstore==0.2
zipp==3.8.0
